code/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openneuro-studies"
version = "0.20251020.0"
description = "Infrastructure for organizing OpenNeuro datasets as BIDS study structures"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Yaroslav O. Halchenko", email = "debian@onerussian.com"}
]
keywords = ["neuroscience", "BIDS", "OpenNeuro", "DataLad", "neuroimaging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "click>=8.1.0",
    "copier>=9.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "requests-cache>=1.1.0",
    "datalad>=0.19.0",
    "PyGithub>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.4.0",
    "types-pyyaml>=6.0.0",
    "types-requests>=2.31.0",
]
sparse = [
    "datalad-fuse>=0.4.0",
    "fsspec>=2023.1.0",
    "aiohttp>=3.8.0",
]
imaging = [
    "nibabel>=5.0.0",
]

[project.scripts]
openneuro-studies = "openneuro_studies.cli.main:cli"

[project.urls]
Homepage = "https://github.com/OpenNeuroStudies/OpenNeuroStudies"
Repository = "https://github.com/OpenNeuroStudies/OpenNeuroStudies"
Issues = "https://github.com/OpenNeuroStudies/OpenNeuroStudies/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["openneuro_studies*", "bids_studies*"]

[tool.setuptools.package-data]
openneuro_studies = ["py.typed", "provision/templates/**/*"]
bids_studies = ["py.typed", "schemas/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=openneuro_studies",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "ai_generated: Tests generated by AI assistants",
    "datalad_install: Tests that perform DataLad recursive install (slow, not run by default)",
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long (handled by black)
]

[tool.ruff.isort]
known-first-party = ["openneuro_studies", "bids_studies"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "datalad.*",
    "datalad_fuse.*",
    "requests_cache.*",
    "fsspec.*",
    "nibabel.*",
    "numpy.*",
]
ignore_missing_imports = true
